Update build process and dependencies
Update the devDependencies to update React and Jest libraries.
With the update to Jest, they deprecated some config options in package.json
such as scriptPrepocessor. Remove testIgnorPatter and scriptPreprocessor and add moduleNameMapper
and unmockedModulePatterns. Add __mocks__ directory and styleMock.js
to be able to mock the scss that gets imported in to the componnet.

Change dev to development and prod to production
in package.json and webpack configs. Add __DEV__ variable
to eslint globals and Jest globals.